+++ source/Energinet.DataHub.MeteringPoints.Application/Connect/ConnectMeteringPoint.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.Common.Transport;
 
-namespace Energinet.DataHub.MeteringPoints.Application
+namespace Energinet.DataHub.MeteringPoints.Application.Connect
 {
     public record ConnectMeteringPoint(
         string GsrnNumber = "",
