+++ source/Energinet.DataHub.MeteringPoints.Application/Connect/ConnectMeteringPointHandler.cs
@@ -23,7 +23,7 @@ using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using MediatR;
 
-namespace Energinet.DataHub.MeteringPoints.Application
+namespace Energinet.DataHub.MeteringPoints.Application.Connect
 {
     public class ConnectMeteringPointHandler : IRequestHandler<ConnectMeteringPoint, BusinessProcessResult>
     {
