+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/XmlConverter/Mappings/ConnectMeteringPointXmlMappingConfiguration.cs
@@ -18,7 +18,7 @@ namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.XmlConverter.Mappi
     {
         public ConnectMeteringPointXmlMappingConfiguration()
         {
-            CreateMapping<Application.ConnectMeteringPoint>("MktActivityRecord", mapper => mapper
+            CreateMapping<Application.Connect.ConnectMeteringPoint>("MktActivityRecord", mapper => mapper
                 .AddProperty(x => x.GsrnNumber, "MarketEvaluationPoint", "mRID")
                 .AddProperty(x => x.TransactionId, "mRID")
                 .AddProperty(x => x.EffectiveDate, "start_DateAndOrTime.dateTime"));
