+++ source/Energinet.DataHub.MeteringPoints.Tests/Application/ConnectMeteringPointMapperTests.cs
@@ -14,6 +14,7 @@
 
 using AutoFixture;
 using Energinet.DataHub.MeteringPoints.Application;
+using Energinet.DataHub.MeteringPoints.Application.Connect;
 using FluentAssertions;
 using Xunit;
 using Xunit.Categories;
